@@ -35,19 +35,21 @@
   "homepage": "https://github.com/richardcyrus/fm-rest-countries-api-app#readme",
   "dependencies": {
     "@heroicons/react": "^1.0.6",
+    "@tanstack/react-query": "^5.90.12",
+    "@tanstack/react-query-devtools": "^5.91.1",
     "prop-types": "^15.8.1",
     "radix-ui": "^1.4.3",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-query": "^3.39.3",
+    "react": "19.2.3",
+    "react-dom": "19.2.3",
     "react-router-dom": "^6.30.2",
     "redaxios": "^0.5.1"
   },
   "devDependencies": {
     "@eslint/compat": "^2.0.0",
     "@eslint/js": "^9.39.2",
-    "@types/react": "^17.0.90",
-    "@types/react-dom": "^17.0.26",
+    "@tanstack/eslint-plugin-query": "^5.91.2",
+    "@types/react": "^19.2.7",
+    "@types/react-dom": "^19.2.3",
     "@vitejs/plugin-react": "^5.1.2",
     "axe-core": "^4.11.0",
     "eslint": "^9.39.2",
